requirements.txt
dlib
pandas
numpy
opencv-python
flask
flask_cors
pillow